requirements.txt
Adafruit-Blinka==8.25.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-dht==4.0.3
adafruit-circuitpython-lis3dh==5.2.1
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.5
Adafruit-PlatformDetect==3.54.0
Adafruit-PureIO==1.1.11
boto3==1.29.3
botocore==1.32.3
isort==5.12.0
jmespath==1.0.1
mypy==1.7.1
mypy-extensions==1.0.0
pyftdi==0.55.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pyusb==1.2.1
RPi.GPIO==0.7.1
ruff==0.1.6
s3transfer==0.7.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
